Update ANtools.ipf
CrawlTree() added. Crawls the waves in all data folders and shows their data size in a summary graph. Allows identification of the most space consuming data.
ColorByGeo now can work with inverted color sequence. 
NormTraces is now aware of the subrange that is plotted. Hence Global Max now means max of the plotted data and the cursor positions are now correctly read out, even if only a range of the wave is plotted.