Fix sign of the phase
Fix sign of the phase as calculated in GainCalculationWrapper and BST_Wrapper: Previously, I used a _Correlation_ (Spike triggered average input) in place of what should have been a convolution. The two differ by the "direction" of time in the kernel. In the frequency that corresponds to inverting the sign pf the phase, i.e. using the conjugated complex of the FFT(STA).