Correct calculation of vector strength's confidence interval
During bootstrapping, previously used the _magnitude_ of vector strength (VS) and then obtained CI for this magnitude. Now each bootstrap average is _complex_ and it is _projected onto the avg vector_ before the magnitude is used. This assures that in cases when the origin (0+i*0) is included in the 95% confidence region, the correct minimal magnitude is negative.
To evoke the old behaviour, switched off the optional parameter:" doComplex=0" when calling the procedure BootstrapVS