Changed normalization of individual STAs
The normalization of trials' autocorrelation and trials' STA does not necessarily have to be identical. This updates makes them identical. Normalization factor is input VARIANCE. Sometimes in the past I have used input SD for STA normalization, because current is present in STA only in first order. However, if the cells' input resistance differs a lot, this normalization biases the weighting.