Refine Xeqt4List with end removal
In Xeqt4List (and the shortcut Xeqt4WList) any occurrence of the token ~ in a command is replaced with each of the strings from a list of strings. And for each instance, the command is executed. 
Sometimes, it is very handy to modify the string from the list before it is used for replacement. Now, it is possible to replace the '~' with the token '§~rmvendX§'. All symbols are fixed, except for the 'X'. This is a stand-in for the letter, number, or underscore that actually defines the first letter of the ending which is being removed. 
When this token is used, the strings from the provided list are put in its place after their ending was removed.