Magnitude Confidence Interval calculation improved
Changed how magnitude confidence interval is calculated: Bootstrapped complex gain values are projected onto the average gain's unit vector before they are evaluated (with respect to magnitude). This makes it possible to have a lower CI boundary that is NEGATIVE. This offers another way to identify for which frequencies the gain is not significantly different from zero. 
Importantly: for frequencies away from the insignificance boundary (where gain crosses noise floor), the confidence interval is changed very little.
Also: applied the change in how the STA peak delay is represented as a phase shift (after Gaussfilter).